Move vertx metrics factory out of Core into ConnectionMetrics class
