+++ xyz-hub-service/src/main/java/com/here/xyz/hub/Core.java
@@ -21,24 +21,13 @@ package com.here.xyz.hub;
 
 import com.here.xyz.hub.util.ConfigDecryptor;
 import com.here.xyz.hub.util.ConfigDecryptor.CryptoException;
-import com.here.xyz.hub.util.metrics.net.ConnectionMetrics.HubHttpClientMetrics;
-import com.here.xyz.hub.util.metrics.net.ConnectionMetrics.HubTCPMetrics;
 import io.vertx.config.ConfigRetriever;
 import io.vertx.config.ConfigRetrieverOptions;
 import io.vertx.config.ConfigStoreOptions;
 import io.vertx.core.Handler;
 import io.vertx.core.Vertx;
 import io.vertx.core.VertxOptions;
-import io.vertx.core.http.HttpClientOptions;
 import io.vertx.core.json.JsonObject;
-import io.vertx.core.metrics.MetricsOptions;
-import io.vertx.core.net.NetClientOptions;
-import io.vertx.core.net.SocketAddress;
-import io.vertx.core.spi.VertxMetricsFactory;
-import io.vertx.core.spi.metrics.ClientMetrics;
-import io.vertx.core.spi.metrics.HttpClientMetrics;
-import io.vertx.core.spi.metrics.TCPMetrics;
-import io.vertx.core.spi.metrics.VertxMetrics;
 import java.io.IOException;
 import java.io.InputStream;
 import java.text.ParseException;
@@ -105,20 +94,16 @@ public class Core {
     final ConfigStoreOptions sysConfig = new ConfigStoreOptions().setType("sys");
     final ConfigRetrieverOptions options = new ConfigRetrieverOptions().addStore(fileStore).addStore(envConfig).addStore(sysConfig).setScanPeriod(24 * 60 * 1000);
 
-    if(vertxOptions == null) {
-      vertxOptions = new VertxOptions()
-              .setWorkerPoolSize(NumberUtils.toInt(System.getenv(Core.VERTX_WORKER_POOL_SIZE), 128))
-              .setPreferNativeTransport(true)
-              .setMetricsOptions(new MetricsOptions()
-                  .setEnabled(true)
-                  .setFactory(createMetricsFactory()));
-      if (debug) {
-        vertxOptions
-                .setBlockedThreadCheckInterval(TimeUnit.MINUTES.toMillis(1))
-                .setMaxEventLoopExecuteTime(TimeUnit.MINUTES.toMillis(1))
-                .setMaxWorkerExecuteTime(TimeUnit.MINUTES.toMillis(1))
-                .setWarningExceptionTime(TimeUnit.MINUTES.toMillis(1));
-      }
+    vertxOptions = (vertxOptions != null ? vertxOptions : new VertxOptions())
+            .setWorkerPoolSize(NumberUtils.toInt(System.getenv(Core.VERTX_WORKER_POOL_SIZE), 128))
+            .setPreferNativeTransport(true);
+
+    if (debug) {
+      vertxOptions
+              .setBlockedThreadCheckInterval(TimeUnit.MINUTES.toMillis(1))
+              .setMaxEventLoopExecuteTime(TimeUnit.MINUTES.toMillis(1))
+              .setMaxWorkerExecuteTime(TimeUnit.MINUTES.toMillis(1))
+              .setWarningExceptionTime(TimeUnit.MINUTES.toMillis(1));
     }
 
     vertx = Vertx.vertx(vertxOptions);
@@ -147,18 +132,6 @@ public class Core {
       initializeLogger(config, debug);
       handler.handle(config);
     });
-  }
-
-  private static VertxMetricsFactory createMetricsFactory() {
-    return new CoreMetricsFactory();
-  }
-
-  private static class CoreMetricsFactory implements VertxMetricsFactory {
-
-    @Override
-    public VertxMetrics metrics(VertxOptions options) {
-      return new CoreMetrics();
-    }
   }
 
   private static void initializeLogger(JsonObject config, boolean debug) {
@@ -226,23 +199,5 @@ public class Core {
     }
 
     return buildProperties.getProperty(name);
-  }
-
-  public static class CoreMetrics implements VertxMetrics {
-
-    @Override
-    public HttpClientMetrics<?, ?, ?, ?> createHttpClientMetrics(HttpClientOptions options) {
-      return new HubHttpClientMetrics();
-    }
-
-    @Override
-    public ClientMetrics<?, ?, ?, ?> createClientMetrics(SocketAddress remoteAddress, String type, String namespace) {
-      return null;
-    }
-
-    @Override
-    public TCPMetrics<?> createNetClientMetrics(NetClientOptions options) {
-      return new HubTCPMetrics();
-    }
   }
 }
