+++ xyz-hub-service/src/main/java/com/here/xyz/hub/Service.java
@@ -19,7 +19,6 @@
 
 package com.here.xyz.hub;
 
-import com.amazonaws.services.cloudwatch.model.StandardUnit;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.here.xyz.hub.auth.Authorization;
@@ -32,20 +31,23 @@ import com.here.xyz.hub.rest.admin.MessageBroker;
 import com.here.xyz.hub.rest.admin.Node;
 import com.here.xyz.hub.rest.admin.messages.RelayedMessage;
 import com.here.xyz.hub.util.ARN;
-import com.here.xyz.hub.util.metrics.base.CWBareValueMetricPublisher;
 import com.here.xyz.hub.util.metrics.GcDurationMetric;
 import com.here.xyz.hub.util.metrics.GlobalInflightRequestMemory;
 import com.here.xyz.hub.util.metrics.GlobalUsedRfcConnections;
 import com.here.xyz.hub.util.metrics.MajorGcCountMetric;
 import com.here.xyz.hub.util.metrics.MemoryMetric;
+import com.here.xyz.hub.util.metrics.base.CWBareValueMetricPublisher;
 import com.here.xyz.hub.util.metrics.base.MetricPublisher;
 import com.here.xyz.hub.util.metrics.net.ConnectionMetrics;
+import com.here.xyz.hub.util.metrics.net.ConnectionMetrics.HubMetricsFactory;
 import io.vertx.core.AsyncResult;
 import io.vertx.core.CompositeFuture;
 import io.vertx.core.DeploymentOptions;
 import io.vertx.core.Future;
 import io.vertx.core.Promise;
+import io.vertx.core.VertxOptions;
 import io.vertx.core.json.JsonObject;
+import io.vertx.core.metrics.MetricsOptions;
 import io.vertx.ext.web.Router;
 import io.vertx.ext.web.client.WebClient;
 import io.vertx.ext.web.client.WebClientOptions;
@@ -140,7 +142,11 @@ public class Service extends Core {
    */
   public static void main(String[] arguments) {
     boolean debug = Arrays.asList(arguments).contains("--debug");
-    initialize(null, debug, CONFIG_FILE, Service::onConfigLoaded);
+    VertxOptions vertxOptions = new VertxOptions()
+        .setMetricsOptions(new MetricsOptions()
+          .setEnabled(true)
+          .setFactory(new HubMetricsFactory()));
+    initialize(vertxOptions, debug, CONFIG_FILE, Service::onConfigLoaded);
   }
 
   /**
